requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai>=1.0.0
faiss-cpu==1.13.1
PyPDF2==3.0.1
numpy==1.26.4
pydantic>=2.0.0
requests>=2.31.0
streamlit==1.32.0
